feat: add user search and error catch
modify index.html with a new input field and
submit button that runs the apiFetch() function
and inserts the search content into the fetch API
URL.

add API error responses to the .then and a newly
added .catch to display error information to
users.

add inline style edits throughout the html to
align the content on the page.